Refactor code
- transfer code to pagination.js
- transfer code to post.js
- transfer code to search.js
- refactor code file home.js
