+++ js/home.js
@@ -1,78 +1,5 @@
 import postApi from './api/postApi'
-import { getUlPagination, setTextContent, truncateText } from './utils'
-import dayjs from 'dayjs'
-import relativeTime from 'dayjs/plugin/relativeTime'
-import debounce from 'lodash.debounce'
-
-// to use fromNow func
-dayjs.extend(relativeTime)
-
-function createPostElement(post) {
-  if (!post) return
-
-  // find and clone template
-  const postTemplate = document.getElementById('postTemplate')
-  if (!postTemplate) return
-
-  const liElement = postTemplate.content.firstElementChild.cloneNode(true)
-  if (!liElement) return
-
-  // update title, description, author, thumbnail
-  setTextContent(liElement, '[data-id="title"]', post.title)
-  setTextContent(liElement, '[data-id="description"]', truncateText(post.description, 100))
-  setTextContent(liElement, '[data-id="author"]', post.author)
-
-  // calculate timespan
-  setTextContent(liElement, '[data-id="timeSpan"]', ` - ${dayjs(post.updatedAt).fromNow()}`)
-
-  const thumbnailElement = liElement.querySelector('[data-id="thumbnail"]')
-  if (thumbnailElement) {
-    thumbnailElement.src = post.imageUrl
-
-    thumbnailElement.addEventListener('error', () => {
-      thumbnailElement.src = 'https://via.placeholder.com/1368x400?text=thumbnail'
-    })
-  }
-
-  // TODO: attach event
-
-  return liElement
-}
-
-function renderPostList(postList) {
-  if (!Array.isArray(postList)) return
-
-  const ulElement = document.getElementById('postList')
-  if (!ulElement) return
-
-  // clear current list
-  ulElement.textContent = ''
-
-  postList.forEach((post, idx) => {
-    const liElement = createPostElement(post)
-    ulElement.appendChild(liElement)
-  })
-}
-
-function renderPagination(pagination) {
-  const ulPagination = getUlPagination()
-  if (!pagination || !ulPagination) return
-  const { _page, _limit, _totalRows } = pagination
-
-  // calc totalPages
-  const totalPages = Math.ceil(_totalRows / _limit)
-
-  // save page and totalPages to ulPagination
-  ulPagination.dataset.page = _page
-  ulPagination.dataset.totalPages = totalPages
-
-  // check if enable/disable prev/next links
-  if (_page <= 1) ulPagination.firstElementChild?.classList.add('disabled')
-  else ulPagination.firstElementChild?.classList.remove('disabled')
-
-  if (_page >= totalPages) ulPagination.lastElementChild?.classList.add('disabled')
-  else ulPagination.lastElementChild?.classList.remove('disabled')
-}
+import { initPagination, initSearch, renderPostList, renderPagination } from './utils'
 
 async function handleFilterChange(filterName, filterValue) {
   try {
@@ -85,101 +12,44 @@ async function handleFilterChange(filterName, filterValue) {
 
     history.pushState({}, '', url)
 
-    // fetch API
-    // re-render post list
+    // fetch API and re-render post list
     const { data, pagination } = await postApi.getAll(url.searchParams)
-    renderPostList(data)
-    renderPagination(pagination)
+    renderPostList('postList', data)
+    renderPagination('pagination', pagination)
   } catch (error) {
     console.log('failed to fetch post list', error)
   }
 }
 
-function handlePrevClick(e) {
-  e.preventDefault()
-
-  const ulPagination = getUlPagination()
-  if (!ulPagination) return
-
-  const page = Number.parseInt(ulPagination.dataset.page) || 1
-  if (page <= 1) return
-
-  handleFilterChange('_page', page - 1)
-}
-
-function handleNextClick(e) {
-  e.preventDefault()
-
-  const ulPagination = getUlPagination()
-  if (!ulPagination) return
-
-  const page = Number.parseInt(ulPagination.dataset.page) || 1
-  const totalPages = ulPagination.dataset.totalPages
-  if (page >= totalPages) return
-
-  handleFilterChange('_page', page + 1)
-}
-
-function initPagination() {
-  // blink click event for prev/next link
-  const ulPagination = getUlPagination()
-  if (!ulPagination) return
-
-  // add click event for prev link
-  const prevLink = ulPagination.firstElementChild?.firstElementChild
-  if (prevLink) {
-    prevLink.addEventListener('click', handlePrevClick)
-  }
-
-  // add click event for next link
-  const nextLink = ulPagination.lastElementChild?.firstElementChild
-  if (nextLink) {
-    nextLink.addEventListener('click', handleNextClick)
-  }
-}
-
-function initURL() {
-  const url = new URL(window.location)
-
-  // update search params if needed
-  if (!url.searchParams.get('_page')) url.searchParams.set('_page', 1)
-  if (!url.searchParams.get('_limit')) url.searchParams.set('_limit', 6)
-
-  history.pushState({}, '', url)
-}
-
-function initSearch() {
-  const searchInput = document.getElementById('searchInput')
-  if (!searchInput) return
-
-  // set default values from query params
-  // title_like
-  const queryParams = new URLSearchParams(window.location.search)
-  if (queryParams.get('title_like')) {
-    searchInput.value = queryParams.get('title_like')
-  }
-
-  const debounceSearch = debounce(
-    (event) => handleFilterChange('title_like', event.target.value),
-    500
-  )
-  searchInput.addEventListener('input', debounceSearch)
-}
-
+// MAIN
 ;(async () => {
   try {
-    // attach click event for links
-    initPagination()
-    initSearch()
-
     // set default pagination (_page, _limit) on URL
-    initURL()
+    const url = new URL(window.location)
+    // update search params if needed
+    if (!url.searchParams.get('_page')) url.searchParams.set('_page', 1)
+    if (!url.searchParams.get('_limit')) url.searchParams.set('_limit', 6)
+
+    history.pushState({}, '', url)
+    const queryParams = url.searchParams
+
+    // attach click event for links
+    initPagination({
+      elementId: 'pagination',
+      defaultParams: queryParams,
+      onChange: (page) => handleFilterChange('_page', page),
+    })
+
+    initSearch({
+      elementId: 'searchInput',
+      defaultParams: queryParams,
+      onChange: (value) => handleFilterChange('title_like', value),
+    })
 
     // render post list based URL params
-    const queryParams = new URLSearchParams(window.location.search)
     const { data, pagination } = await postApi.getAll(queryParams)
-    renderPostList(data)
-    renderPagination(pagination)
+    renderPostList('postList', data)
+    renderPagination('pagination', pagination)
   } catch (error) {
     console.log('get all failed', error)
     // show modal, toast error
